2nd commit

+++ server.py
@@ -1,59 +1,64 @@
 #!/usr/bin/env python3
 
-import socket, sys, time, os, threading, select
+# startup
+if True:
+	import socket, sys, time, os, threading, select
 
-#=====================
-PORT = 8761
-IP = socket.gethostbyname(socket.gethostname())
-# 127.0.1.1
-HOST = socket.gethostname()
-HEADER_LENGTH = 64
-FORMAT = 'utf-8'
-#=====================
+	#=====================
+	PORT = 8761
+	IP = socket.gethostbyname(socket.gethostname())
+	# 127.0.1.1
+	HOST = socket.gethostname()
+	HEADER_LENGTH = 64
+	FORMAT = 'utf-8'
+	EXITCOMMAND = 'quit'
+	#=====================
 
-# server startup
-print('[SERVER] server startup...')
-server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-server.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+	# server startup
+	print('[SERVER] server startup...')
+	server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+	server.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
 
-try:
-	server.bind((IP, PORT))
-	print(f"[SERVER] Server bound. IP: {IP} port: {PORT}")
-except OSError:
-	print(f"[SERVER] Could not bind. Aborting...\n")
-	server.close()
-	os.system('pkill -f ".py"')
+	try:
+		server.bind((IP, PORT))
+		print(f"[SERVER] Server bound. IP: {IP} port: {PORT}")
+	except OSError:
+		print(f"[SERVER] Could not bind. Aborting...\n")
+		server.close()
+		os.system('pkill -f ".py"')
 
-server.listen()
-print(f"[SERVER] Listening for connections...")
-sockets_list = [server]
-clients = {}
+	server.listen()
+	print(f"[SERVER] Listening for connections...")
+	sockets_list = [server]
+	sockets_write_list = []
+	clients = {}
 
 def receive_message(client_socket):
-	try:
-		# receive header
-		message_header = client_socket.recv(HEADER_LENGHT)
+    try:
+        # Receive our "header" containing message length, it's size is defined and constant
+        message_header = client_socket.recv(HEADER_LENGTH)
 
-		if not len(message_header):
-			# connection closed
-			return False
+        # If we received no data, client gracefully closed a connection, for example using socket.close() or socket.shutdown(socket.SHUT_RDWR)
+        if not len(message_header):
+            return False
 
-		# get message lenght and receive it
-		message_length = int(message_header.decode(FORMAT).strip())
-		return {'header': message_header, 'content': client_socket.recv(message_length)}
+        # Convert header to int value
+        message_length = int(message_header.decode(FORMAT).strip())
 
-	except:
-		return False
+        # Return an object of message header and message data
+        return {'header': message_header, 'content': client_socket.recv(message_length)}
+
+    except:
+        return False
 
 while True:
-	read_sockets, _, exception_sockets = select.select(sockets_list, [], sockets_list)
+	read_sockets, write_sockets, exception_sockets = select.select(sockets_list, sockets_write_list, sockets_list)
 
 	for notified_socket in read_sockets:
 		if notified_socket == server:
 			# someone connected
 			client_socket, client_address = server.accept()
 			user = receive_message(client_socket)
-
 			if user is False:
 				# someone disconnected
 				continue
@@ -63,7 +68,7 @@ while True:
 
 			info1 = f"{client_address[0]}:{client_address[1]}"
 			info2 = f"{user['content'].decode(FORMAT)}"
-			print(f'New connection from {info1} username: {info2}')
+			print(f'[SERVER] New connection from {info1} username: {info2}')
 
 		else:
 			# received message
@@ -71,13 +76,13 @@ while True:
 
 			if message is False:
 				# client closed connection
-				print(f"Closed connection from {clients[notified_socket]['content'].decode(FORMAT)}")
+				print(f"[SERVER] Closed connection from {clients[notified_socket]['content'].decode(FORMAT)}")
 				sockets_list.remove(notified_socket)
 				del clients[notified_socket]
 				continue
 
 			user = clients[notified_socket]
-			print(f"Received message from {user['content'].decode(FORMAT)}: '{message['content'].decode(FORMAT)}'")
+			print(f"[SERVER] Received message from {user['content'].decode(FORMAT)}: '{message['content'].decode(FORMAT)}'")
 
 			# spread message
 			for client_socket in clients:
@@ -87,3 +92,4 @@ while True:
 	for notified_socket in exception_sockets:
 		sockets_list.remove(notified_socket)
 		del clients[notified_socket]
+
